[ci] write test suites to rockset
Currently we upload all `testcase` elements as individual test runs to
Rockset. It would be nice to also have `testsuite`s as well, which
aggregate high level information.

These aggregations could technically be performed in the backend, but it's
faster to just log the data since we already have it in the XML test
report.

ghstack-source-id: 3dbfa617e634a2e9fb6f1c3978b45c851f3239fa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79265