[quant] Signatures check for F.convXd
Introducing tests to make sure the signatures of functionals match
between quantized and FP. One of the issues in the past was that the
quantized functionals had arguments that were not included in the
FP counterpart of it. For example, the conv1d used to have `padding_mode`
which is not needed in the nn.functional.conv1d, but was present in the
nn.quantized.functional.conv1d.

Test Plan:

```python
python test/test_quantization.py TestSignatureEquivalence
```

ghstack-source-id: 9b9a2bae5c6960ae0251ab96cae7178ce24327c9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/78543